change data_type_serde to fix orc_reader compiler check

+++ be/src/vec/data_types/serde/data_type_bitmap_serde.cpp
@@ -60,7 +60,7 @@ Status DataTypeBitMapSerDe::serialize_one_cell_to_json(const IColumn& column, in
 }
 
 Status DataTypeBitMapSerDe::deserialize_column_from_json_vector(
-        IColumn& column, std::vector<Slice>& slices, int* num_deserialized,
+        IColumn& column, std::vector<Slice>& slices, uint64_t* num_deserialized,
         const FormatOptions& options) const {
     DESERIALIZE_COLUMN_FROM_JSON_VECTOR()
     return Status::OK();
