+++ be/src/vec/data_types/serde/data_type_datetimev2_serde.h
@@ -56,7 +56,7 @@ public:
                                           const FormatOptions& options) const override;
 
     Status deserialize_column_from_json_vector(IColumn& column, std::vector<Slice>& slices,
-                                               int* num_deserialized,
+                                               uint64_t* num_deserialized,
                                                const FormatOptions& options) const override;
 
     void write_column_to_arrow(const IColumn& column, const NullMap* null_map,
@@ -77,10 +77,10 @@ public:
                                int64_t start, int64_t end,
                                std::vector<StringRef>& buffer_list) const override;
 
-    Status deserialize_column_from_fixed_json(IColumn& column, Slice& slice, int rows,
-                                              int* num_deserialized,
+    Status deserialize_column_from_fixed_json(IColumn& column, Slice& slice, uint64_t rows,
+                                              uint64_t* num_deserialized,
                                               const FormatOptions& options) const override;
-    void insert_column_last_value_multiple_times(IColumn& column, int times) const override;
+    void insert_column_last_value_multiple_times(IColumn& column, uint64_t times) const override;
 
 private:
     template <bool is_binary_format>
