+++ be/src/vec/data_types/serde/data_type_jsonb_serde.cpp
@@ -102,7 +102,7 @@ Status DataTypeJsonbSerDe::serialize_one_cell_to_json(const IColumn& column, int
 
 Status DataTypeJsonbSerDe::deserialize_column_from_json_vector(IColumn& column,
                                                                std::vector<Slice>& slices,
-                                                               int* num_deserialized,
+                                                               uint64_t* num_deserialized,
                                                                const FormatOptions& options) const {
     DESERIALIZE_COLUMN_FROM_JSON_VECTOR();
     return Status::OK();
