+++ be/src/vec/data_types/serde/data_type_map_serde.cpp
@@ -136,7 +136,7 @@ Status DataTypeMapSerDe::deserialize_one_cell_from_hive_text(
 }
 
 Status DataTypeMapSerDe::deserialize_column_from_hive_text_vector(
-        IColumn& column, std::vector<Slice>& slices, int* num_deserialized,
+        IColumn& column, std::vector<Slice>& slices, uint64_t* num_deserialized,
         const FormatOptions& options, int hive_text_complex_type_delimiter_level) const {
     DESERIALIZE_COLUMN_FROM_HIVE_TEXT_VECTOR();
     return Status::OK();
@@ -178,7 +178,7 @@ Status DataTypeMapSerDe::serialize_one_cell_to_hive_text(
 
 Status DataTypeMapSerDe::deserialize_column_from_json_vector(IColumn& column,
                                                              std::vector<Slice>& slices,
-                                                             int* num_deserialized,
+                                                             uint64_t* num_deserialized,
                                                              const FormatOptions& options) const {
     DESERIALIZE_COLUMN_FROM_JSON_VECTOR()
     return Status::OK();
