+++ be/src/vec/data_types/serde/data_type_number_serde.h
@@ -67,14 +67,14 @@ public:
                                           const FormatOptions& options) const override;
 
     Status deserialize_column_from_json_vector(IColumn& column, std::vector<Slice>& slices,
-                                               int* num_deserialized,
+                                               uint64_t* num_deserialized,
                                                const FormatOptions& options) const override;
 
-    Status deserialize_column_from_fixed_json(IColumn& column, Slice& slice, int rows,
-                                              int* num_deserialized,
+    Status deserialize_column_from_fixed_json(IColumn& column, Slice& slice, uint64_t rows,
+                                              uint64_t* num_deserialized,
                                               const FormatOptions& options) const override;
 
-    void insert_column_last_value_multiple_times(IColumn& column, int times) const override;
+    void insert_column_last_value_multiple_times(IColumn& column, uint64_t times) const override;
 
     Status write_column_to_pb(const IColumn& column, PValues& result, int64_t start,
                               int64_t end) const override;
